remove redundant test & file
