+++ pkg/addons/default/addons.go
@@ -2,7 +2,9 @@ package defaultaddons
 
 import (
 	"github.com/aws/aws-sdk-go/service/eks/eksiface"
+	"github.com/pkg/errors"
 	"github.com/weaveworks/eksctl/pkg/kubernetes"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type AddonInput struct {
@@ -44,4 +46,13 @@ func DoAddonsSupportMultiArch(eksAPI eksiface.EKSAPI, rawClient kubernetes.RawCl
 		return true, err
 	}
 	return coreDNSUpToDate, nil
+}
+
+// LoadAsset return embedded manifest as a runtime.Object
+func newList(data []byte) (*metav1.List, error) {
+	list, err := kubernetes.NewList(data)
+	if err != nil {
+		return nil, errors.Wrapf(err, "loading individual resources from manifest")
+	}
+	return list, nil
 }
